adding additional figure and figure caption
