+++ manuscript/03_results.md
@@ -24,10 +24,12 @@ some introduction...
 
 ### Performance based on varying distribution similarity
 
-![**Performance based on varying distribution similarity.** Single uncensored lineages with fate and time observations are simulated with two states. These lineages are created with ${\pi=\begin{pmatrix}{0.5 \\ 0.5}\end{pmatrix}$ and ${T=\begin{pmatrix}{0 & 1 \\ 1 & 0}\end{pmatrix}$ with $2^{12}-1$ cells to ensure that cells from each state are well represented in the lineage. One state is fixed and the other is allowed to vary in terms of its distribution. Initially, the two states are the same, i.e., they generate exactly the same distribution of cells. The second state is then allowed to vary by increasing the gamma shape parameter. Doing so increases the Wassertein distance between the two states. $750$ cells of each state from each lineage are recorded and the sampled distributions are used to calculate the Wasserstein distance between the states. The Bernoulli parameter (a) and the gamma scale parameter (c) are the same for both of the states, but the gamma shape (b) parameter varies for the second lineage. The model is able to estimate the parameters of the states across all Wasserstein distances but has difficulty when the two states are identical, when the Wasserstein distance between the states is $0$.](./output/figure7.svg){#fig:wass}
+![**Performance based on varying distribution similarity with uncensored lineages.** Single uncensored lineages with fate and time observations are simulated with two states. These lineages are created with ${\pi=\begin{pmatrix}{0.5 \\ 0.5}\end{pmatrix}$ and ${T=\begin{pmatrix}{0 & 1 \\ 1 & 0}\end{pmatrix}$ with $2^{12}-1$ cells to ensure that cells from each state are well represented in the lineage. One state is fixed and the other is allowed to vary in terms of its distribution. Initially, the two states are the same, i.e., they generate exactly the same distribution of cells. The second state is then allowed to vary by increasing the gamma shape parameter. Doing so increases the Wassertein distance between the two states. Cells of each state from each lineage are sampled as distributions and are used to calculate the Wasserstein distance between the states. The Bernoulli parameter (a) and the gamma scale parameter (c) are the same for both of the states, but the gamma shape (b) parameter varies for the second lineage. The model is able to estimate the parameters of the states across all Wasserstein distances but has difficulty when the two states are identical, when the Wasserstein distance between the states is $0$.](./output/figure7.svg){#fig:wass1}
+
+![**Performance based on varying distribution similarity with censored lineages.** Single censored lineages with fate and time observations are simulated with two states. These lineages are created with ${\pi=\begin{pmatrix}{0.5 \\ 0.5}\end{pmatrix}$ and ${T=\begin{pmatrix}{0 & 1 \\ 1 & 0}\end{pmatrix}$ with $2^{12}-1$ cells to ensure that cells from each state are well represented in the lineage just as before. These lineages are then censored; cells that die or live longer than the maximum experimental time have their subtrees removed. As before, one state is fixed and the other is allowed to vary in terms of its distribution. The two states are initally the same, i.e., they generate exactly the same distribution of cells. The second state is then allowed to vary by increasing the gamma shape parameter. Doing so increases the Wassertein distance between the two states. Cells of each state from each lineage are sampled as distributions and are used to calculate the Wasserstein distance between the states. The Bernoulli parameter (a) and the gamma scale parameter (c) are the same for both of the states, but the gamma shape (b) parameter varies for the second lineage. The model is able to estimate the parameters of the states across all Wasserstein distances but has difficulty when the two states are identical, when the Wasserstein distance between the states is $0$.](./output/figure8.svg){#fig:wass1}
 
 
 ### AIC Figure
 
 some introduction...
-![Model prediction accuracy on a simulated cell population with varying numbers of lineages. Accuracy of tHMM state assignment (left), Bernoulli parameter estimation (middle), and exponential parameter estimation (right) for populations of cells with different numbers of lineages. This is analogous to increasing the number of initial seeded cells. Lineages were composed of initially therapy sensitive (blue) that then later transitioned to resistant (orange) subpopulations. Dotted horizontal lines at target accuracy and true population values are depicted. Solid line represents a 10-point moving average. ...](./output/figure8.svg){#fig:AIC}
+![Model prediction accuracy on a simulated cell population with varying numbers of lineages. Accuracy of tHMM state assignment (left), Bernoulli parameter estimation (middle), and exponential parameter estimation (right) for populations of cells with different numbers of lineages. This is analogous to increasing the number of initial seeded cells. Lineages were composed of initially therapy sensitive (blue) that then later transitioned to resistant (orange) subpopulations. Dotted horizontal lines at target accuracy and true population values are depicted. Solid line represents a 10-point moving average. ...](./output/figure9.svg){#fig:AIC}
